Refactor dynamic decimal places formatter
